+++ CHANGELOG.md
@@ -32,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Restrcutured all Library articles
+- Restructured all Library articles
 
 ## [0.1.0] - 2026-04-23